app/facade/requirements.txt
fastapi
httpx
python-consul
